// src/app/canvas/types/endboss.type.ts
export type EndBoss = {
  live: number;
  defeatedAt: number;
  lastHitTakenAt: number;
  lastWalkAnimationAt: number;
  lastHitAnimationAt: number;
  status: string;
  deathImgNr: number;
  walkImgNr: number;
  hurtImgNr: number;
  alertImgNr: number;
  attackImgNr: number;
  moveLeft: boolean;
  imgSrc: string;
  pos_x: number;
  pos_y: number;
};